Restart home controls activity when dream is redirecting wakes
When the dream is redirecting calls to wakeUp, the dream is not
terminated but the embedded activity is. This results in a blank dream.
The fix here will restart the embedded activity in this case, after a
short delay to allow the wakeup animation to play first.

Fixes: 341957747
Test: atest HomeControlsDreamServiceTest
Flag: android.service.controls.flags.home_panel_dream
Change-Id: I495b20235affa49e31b14bc0e50acee007f2813d