Use uinput-based injection in AnrTest
Prior to this CL, we were injecting events into the unresponsive
activity using "injectInputEvent". The up event was missing, which would
cause the pointer to linger in the gesture monitors. Later in
the test, we were clicking the "close" button. This click would start to
interact with the previous unreleased pointer, causing the notification
shade to get pulled down.

In this CL, we switch to uinput-based injection to make it hermetic.

Bug: 343912075
Test: atest com.android.test.input.AnrTest
Change-Id: I295963080203ee64f50ed9890086b945665c204e