Fix untrusted activity not shown when pinned
Before, we only check if the activity on top of untrusted activity
is in different UID and non-finishing, but it doesn't cover the case
that the untrusted activity is in the primary split container, while
the top activity of the secondary split is in the same process of
the host task.

This CL also checks if the activity above the untrusted activity
occludes the untrusted activity to fix the issue.

Test: atest TaskFragmentTest
Fixes: 343308303
Flag: EXEMPT bugfix
Change-Id: I8a94207f4b5a69aadcffa13266e32fd18d54b150