Turn off the kill on cached idle and bg restricted processes
The previous CL a294891553a15 was designed to kill processes
which is in the cached state, being idle, and also bg restricted,
in order to minimize the impact on the sys health. Now given
we have freezer in place by default, that behavior isn't necessary
anymore. Turn if off for now.

Bug: 346090221
Test: Manual - Start a BG restricted FGS, background it and verify it's not killed.
Change-Id: Id0f07c8bdb9d65ffedcec74373fbdeaefc3b8f9a