Load aconfig demo flags in system partition
Aconfig demo flags do not appear in CLI tooling, because they are not
loaded into the system partition on build, since they're not required
anywhere except the demo app which is not preinstalled. Add them as a
dependency to SettingsProviderLib, reconciling with how internal main is
setup.

Test: m
Bug: 346839125
Change-Id: Ibce71b0047a9b44575460b36adba6425566a6299
Merged-In: I81c643ce0121b0718692fd23f5b06a2b641f2b1e