Shell cmd power-reset instead of power-on
power-on command may not work as expected,
when display already asleep. The semantics
of power-off command means we need a
power-reset command to reset the display
to a state it supposed to have by removing
the 'overridden' power state.

Change-Id: Ib7acfa6cb7687df4f19877e3b9cf9b8d768151d6
Bug: 292300981
Test: atest DisplayManagerServiceTest
Test: adb shell cmd display power-off 0
Test: adb shell cmd display power-reset 0
Flag: EXEMPT bugfix in a new shell command not used anywhere