Updates containers if the updates skipped when Task was invisible
The activity in the secondary container was removed and the
primary container was not updated because the the Task was
invisible. The next time the Task was brought to front, the
primary container still not being updated, e.g. expands the
primary container or starts placeholder on the secondary
container, because the Task was not resized (commit eee645e).

Bug: 344721335
Test: wm presubmit
Test: repo steps on the bug
Flag: com.android.window.flags.fix_no_container_update_without_resize
Change-Id: I0165df4ef65bd08e6f53b4323816dc8dcfa4e75a