Expose IMMI object for Proxy class
Expose IMMI object from IMMS so that IMMSProxy can redirect the internal
call based on user to respective IMMI.

IMMSProxy is a temporary plan. Once the long-term solution for concurrent multi-user IME is implemented, this change will be reverted since this IMMI ref is not going to be used anymore.

Flag: NONE <Exposing package scoped method>
Bug: 350790759
Test: m

Change-Id: Ia3cfcef9c5d176c74dff8120e2474263e3e73a9f