Remove getCurrentInputMethodSubtypeForNonCurrentUsers()
This is a follow up CL to my previous CL [1], which enabled background
users to call

  InputMethodManager#getCurrentInputMethodSubtype()

by introducing

  #getCurrentInputMethodSubtypeForNonCurrentUsers().

The reason why we could not simply reuse the same implementation for

  InputMethodManagerService#getCurrentInputMethodSubtypeLocked()

was due to the fact that IMMS#mCurrentSubtype was not per-user.

Now that #mCurrentSubtype is fully per-user [2], we should be able to
unify the code.

 [1]: Icb09f9cb1a4147884faa9952b2e03ec4afa9f0b1
      b8c05216d13f8d780c51dd401e8fd18ff0c241ed
 [2]: Ibcc39212ead12c2897f8f52cb9f15b1d39959f71
      dac2fe914a31a0770f4e4c54ad754779eb35f31f

Bug: 325515685
Bug: 350386877
Test: presubmit
Flag: EXEMPT refactor
Change-Id: Ia52b556ff3df3d2b50b9d62198455e5ce6ebe659