Maintain fixed-rotation illusion for wallpapers in transition
Wallpaper tokens, like activities, can be nested in a fixed
rotation leash to abstract it away from shell. So, this adds
the same logic as used for activity to the wallpapertoken.

Additionally, shell only receives the rotation-leash, so it
can't modify the token's visibility -- so this also ensures
that the token is made visible.

Bug: 349966512
Flag: com.android.window.flags.ensure_wallpaper_in_transitions
Test: close any fixed-rotation task into launcher.
Change-Id: Ic745390edf25ef2e5f7172a8cc4e1dde09875514