multiPointerDraggable shouldn't start drag gesture with consumed events
ag/28051296 calculates startedPosition when at least one pointer is
present on the screen.
Previously, we used awaitConsumableEvent to determine the starting
position.

The drag gesture might begin with start followed by an immediate
cancellation if the initial position turns out to be a consumed event.

This behavior is now fixed with this CL, as onDragStart requires that
the event is consumable.

Test: atest MultiPointerDraggableTest
Bug: 330200163
Flag: com.android.systemui.scene_container
Change-Id: If9f58c3774bdc26c0dae104a87d26b0541b4e4a0