Make sure to call SingleThreadedBackgroundWriter#onUserCreated() again
This is a follow up CL to my previous CL, which aimed to remove
UserLifecycleListener dependencies from random utility classes under

  com.android.server.inputmethod

package.

One thing I forgot to update was that

  AdditionalSubtypeMapRepository
      .SingleThreadedBackgroundWriter#onUserCreated()

stopped being called after that CL.

With this CL, we start calling it again as it used to be.

 [1]: I337fb084a9395eed7422d6889bf4f15d583c769c
      91bfe7d63b385687a5f856fe077f2698779f6ef8

Bug: 352354308
Test: presubmit
Test: atest FrameworksInputMethodSystemServerTests
Flag: EXEMPT refactor
Change-Id: I69786e5d22a4ef92ab52dd0b8304e40b87f1c9b5