Unify IMMS#onSecureSettingsChangedLocked() behaviors among users
This is a follow up CL to my previous CLs [1], which made our secure
settings observer multi-user aware.

What is remaining is an early exit case at the top of

  InputMethodManagerService#onSecureSettingsChangedLocked(),

which basically makes the above callback no-op when called for
background users when IMMS#mConcurrentMultiUserModeEnabled is false.

With this CL the above early exit rule is removed, meaning that
updates on secure settings will be propagated into downstream logic
even if called for background users.

 [1]: I35253d3964f37f3e13888c0017a86449cf1de8ac
      aec13ae557b9e9e2e90a058fd6b014c0af2dc613
 [2]: I27f94ca795f3035b5132a28fbbec669122d42b87
      12fb4f47de923d38d965deb2b34fd5a07b494fe8

Bug: 342027196
Bug: 350386877
Flag: android.view.inputmethod.concurrent_input_methods
Test: presubmit
Change-Id: I039816fa3e2acfe3cada3fb237c01560d474aee8