Initiate drag and drop using channel tokens
When the drag and drop is started, the input system only needs to know
the tokens of the windows that are involved. The input channels don't
need to be provided; in fact, in a future CL, the input channel won't
even be stored inside the caller (since the WM code initiates drag and
drop on behalf of the app).

In this CL, only the source is modified to use the channel token instead
of channel. In a separate CL, the target will also be modified.

Bug: 323450804
Test: atest android.view.surfacecontrol.cts.SurfaceControlViewHostTests
Test: atest WmTests:com.android.server.wm.DragDropControllerTests
Flag: EXEMPT refactor
Change-Id: Ie1d4a1a5d0be1c2b80d33a160e1b00ec69fe2517