Avoid the caption bar overlaps with status bar
When a window under desktop mode, and get fullscreen when the maximize
button on the caption bar is pressed, the caption bar will layout
starting from (0, 0) point of the display, and it will overlaps with the
status bar, made the buttons on caption bar hard to interact with.

This patch extend the caption bar by the status bar size when it
overlaps and maximized, also added padding to the corresponding view to
make the buttons on the caption moving downward to be accessible by the
users.

The extension only applied to CaptionWindowDecoration.

Test: WindowDecorationTests
Test: Check desktop mode targets
Bug: 351920380
Flag: EXEMPT bugfix
Change-Id: I0065cdc3e51be187d24e617fb97f191a69c86a3c