Add support for mouse keys on non-QWERTY keyboard layouts
Depending on the keyboard layout in use, the mouse keys key codes should
change such that the position of the keys stays the same.
The default is to return the key codes for the
default layout (QWERTY), otherwise return the key code of the given
location on a reference QWERTY keyboard layout.
Also, the user could change the keyboard layout from settings at any
time, which would change the key character map for that device.
Hence, the key code to enum mapping should be updated every time a change
in keyboard is detected. To ensure thread safety all access to the enum
mapping for a particular device should happen from the same thread,
i.e., the handler thread.

bug: 341799888
Test: Manual testing on multiple QWERTY and DVORAK keyboard layouts,
      atest FrameworksServicesTests:MouseKeysInterceptorTest
Flag: EXEMPT bugfix
Change-Id: I4dbea8d4def95f7a710dd3b85e46d214f05fc94f