[SB] Don't call SBWindowController#refreshSBHeight from PhoneSBView.
The `#refreshStatusBarHeight` call was added to address b/323299264,
but if I remove the call, I'm still unable to repro b/323299264. My
hunch is that b/323299264 was fixed by a revert in some other part of
the codebase, so the `#refreshStatusBarHeight` call is unnecessary.
But, I'm protecting the removal of that method call behind a flag so
that we can verify this hunch in stages before sending it to everyone.

Bug: 360115167
Flag: com.android.systemui.status_bar_stop_updating_window_height
Test: On unfolded, start in portrait, then rotate clockwise to 90deg,
then rotate clockwise to 270deg -> verify status bar does not get cut
off
Test: atest PhoneStatusBarViewTest

Change-Id: I85b2f4db944232e3fc5aeedbb8268ea41735af79