Make rule updates "by user in app" more powerful
User actions in app should be capable of removing previous "user action from systemui" overrides.

However, they do not add new overrides. The rationale is that an override in systemui is possibly not known to the app (and certainly unknown unless they listen to the broadcasts added in VIC) but a user action inside the app _is_ known, so it would be strange to disregard fiurther state updates from the app until it "agrees" with itself.

Fixes: 360794520
Test: atest ZenModeHelperTest
Flag: android.app.modes_ui
Change-Id: If86a13f09d3273f4caa83808701252e313dab788